+++ integration/compactor_test.go
@@ -20,6 +20,7 @@ import (
 	"github.com/oklog/ulid"
 	"github.com/prometheus/prometheus/model/histogram"
 	"github.com/prometheus/prometheus/model/labels"
+	"github.com/prometheus/prometheus/tsdb"
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
 	"github.com/prometheus/prometheus/tsdb/chunks"
 	"github.com/prometheus/prometheus/tsdb/index"
@@ -73,16 +74,16 @@ func TestCompactBlocksContainingNativeHistograms(t *testing.T) {
 			Labels: labels.FromStrings("case", "native_histogram", "i", strconv.Itoa(i)),
 			Chunks: []chunks.Meta{
 				tsdbutil.ChunkFromSamples([]tsdbutil.Sample{
-					newSample(10, 0, generateTestHistogram(1), nil),
-					newSample(20, 0, generateTestHistogram(2), nil),
+					sample{10, 0, tsdb.GenerateTestHistogram(1), nil},
+					sample{20, 0, tsdb.GenerateTestHistogram(2), nil},
 				}),
 				tsdbutil.ChunkFromSamples([]tsdbutil.Sample{
-					newSample(30, 0, generateTestHistogram(3), nil),
-					newSample(40, 0, generateTestHistogram(4), nil),
+					sample{30, 0, tsdb.GenerateTestHistogram(3), nil},
+					sample{40, 0, tsdb.GenerateTestHistogram(4), nil},
 				}),
 				tsdbutil.ChunkFromSamples([]tsdbutil.Sample{
-					newSample(50, 0, generateTestHistogram(5), nil),
-					newSample(2*time.Hour.Milliseconds()-1, 0, generateTestHistogram(6), nil),
+					sample{50, 0, tsdb.GenerateTestHistogram(5), nil},
+					sample{2*time.Hour.Milliseconds() - 1, 0, tsdb.GenerateTestHistogram(6), nil},
 				}),
 			},
 		}
@@ -215,9 +216,6 @@ type sample struct {
 	fh *histogram.FloatHistogram
 }
 
-func newSample(t int64, v float64, h *histogram.Histogram, fh *histogram.FloatHistogram) tsdbutil.Sample {
-	return sample{t, v, h, fh}
-}
 func (s sample) T() int64                      { return s.t }
 func (s sample) V() float64                    { return s.v }
 func (s sample) H() *histogram.Histogram       { return s.h }